Rename WorkflowEditor component for consistency
