+++ client/galaxy/scripts/components/Workflow/Editor/Index.vue
